requirements.txt
slackclient
scholarly
pyyaml
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
